requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
telethon==1.34.0
APScheduler>=3.10,<4.0
tzlocal>=3.0
